refactor state of form submission from NewProject into Body * body passed props about form values into NewProject * body is getting a bit messy
